👌 Fix lint errors
